fix test_until_time failure when run early
Since introduction of the "until" job scheduling method to apply
deadlines, the test suite fails at test_until_time when run between
midnight and 5am local hour.  This patch ought to ensure the deadline
is always one hour ago independently of the time of the day, or night.
This should resolve Github issue #488.

Signed-off-by: Étienne Mollier <emollier@debian.org>